+++ src/formula/function.cpp
@@ -16,13 +16,11 @@
 #include "formula/function.hpp"
 
 #include "color.hpp"
-#include "filesystem.hpp"
 #include "formula/callable_objects.hpp"
 #include "formula/debugger.hpp"
 #include "game_config.hpp"
 #include "game_display.hpp"
 #include "log.hpp"
-#include "map/label.hpp"
 #include "pathutils.hpp"
 #include "serialization/unicode.hpp"
 
@@ -267,21 +265,6 @@ DEFINE_WFL_FUNCTION(debug_float, 2, 3)
 		display_float(location, text);
 		return var2;
 	}
-}
-
-DEFINE_WFL_FUNCTION(debug_label, 2, 2)
-{
-	variant var0 = args()[0]->evaluate(variables, fdb);
-	variant var1 = args()[1]->evaluate(variables, fdb);
-
-	if(game_config::debug && display::get_singleton()) {
-		const map_location loc = var0.convert_to<location_callable>()->loc();
-		const std::string text = var1.is_string() ? var1.as_string() : var1.to_debug_string();
-
-		display::get_singleton()->labels().set_label(loc, text);
-	}
-
-	return variant(std::vector<variant>{var0, var1});
 }
 
 DEFINE_WFL_FUNCTION(debug_print, 1, 2)
@@ -1358,7 +1341,7 @@ DEFINE_WFL_FUNCTION(nearest_loc, 2, 2)
 	});
 #endif
 
-	if(nearest == locations.end()) {
+	if(nearest != locations.end()) {
 		return variant(std::make_shared<location_callable>(nearest->convert_to<location_callable>()->loc()));
 	} else {
 		return variant();
@@ -1473,25 +1456,6 @@ DEFINE_WFL_FUNCTION(rotate_loc_around, 2, 3)
 		: 1;
 
 	return variant(std::make_shared<location_callable>(loc.rotate_right_around_center(center, n)));
-}
-
-DEFINE_WFL_FUNCTION(run_file, 1, 1)
-{
-	const std::string file = args()[0]
-		->evaluate(variables, add_debug_info(fdb, 0, "run_file:file"))
-		.string_cast();
-
-	// NOTE: get_wml_location also filters file path to ensure it doesn't contain things like "../../top/secret"
-	auto path = filesystem::get_wml_location(file);
-	if(!path) {
-		return variant();
-	}
-
-	auto builtins = function_symbol_table::get_builtins();
-	std::string formula_string = filesystem::read_file(path.value());
-
-	auto parsed_formula = formula::create_optional_formula(formula_string, builtins.get());
-	return formula::evaluate(parsed_formula, variables, add_debug_info(fdb, -1, "run_file:formula_from_file"));
 }
 
 DEFINE_WFL_FUNCTION(type, 1, 1)
@@ -1671,7 +1635,6 @@ std::shared_ptr<function_symbol_table> function_symbol_table::get_builtins()
 		DECLARE_WFL_FUNCTION(max);
 		DECLARE_WFL_FUNCTION(choose);
 		DECLARE_WFL_FUNCTION(debug_float);
-		DECLARE_WFL_FUNCTION(debug_label);
 		DECLARE_WFL_FUNCTION(debug_print);
 		DECLARE_WFL_FUNCTION(debug_profile);
 		DECLARE_WFL_FUNCTION(wave);
@@ -1707,7 +1670,6 @@ std::shared_ptr<function_symbol_table> function_symbol_table::get_builtins()
 		DECLARE_WFL_FUNCTION(relative_dir);
 		DECLARE_WFL_FUNCTION(direction_from);
 		DECLARE_WFL_FUNCTION(rotate_loc_around);
-		DECLARE_WFL_FUNCTION(run_file);
 		DECLARE_WFL_FUNCTION(index_of);
 		DECLARE_WFL_FUNCTION(keys);
 		DECLARE_WFL_FUNCTION(values);
